Move axis rendering code from envelope to a separate module.
